+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `client_from_env` was moved from `teloxide::utils` to crate root of `teloxide-core`
   - To simplify `GetUpdates` request it was changed to simply return `Vec<Update>` 
     (instead of `Vec<Result<Update, (Value, serde_json::Error)>>`)
+- `GetUpdatesNonStrict` 'telegram' method, that behaves just like `GetUpdates` but doesn't 
+  fail if one of updates fails to be deserialized 
 
 ### Changed